requirements-dev.txt
pytest
black
flake8
pre-commit
pytest-selenium
coverage
tox
